big intenger

+++ app/auth/models/oauth.py
@@ -1,7 +1,7 @@
 from enum import Enum
 from typing import TYPE_CHECKING
 
-from sqlalchemy import Boolean, Enum as SQLEnum, ForeignKey, Integer, String
+from sqlalchemy import BigInteger, Boolean, Enum as SQLEnum, ForeignKey, String
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from app.core.db.base_model import BaseModel, DateMixin
@@ -19,7 +19,7 @@ class OAuthProviderEnum(Enum):
 class OAuthAccount(BaseModel, DateMixin):
     __tablename__ = "oauth_accounts"
 
-    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
+    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
 
     provider: Mapped[OAuthProviderEnum] = mapped_column(SQLEnum(OAuthProviderEnum), nullable=False)
     provider_user_id: Mapped[str] = mapped_column(String, nullable=False)
@@ -27,6 +27,6 @@ class OAuthAccount(BaseModel, DateMixin):
 
     is_active: Mapped[bool] = mapped_column(Boolean, default=True)
 
-    user_id: Mapped[int] = mapped_column(Integer, ForeignKey("users.id", ondelete="cascade", onupdate="cascade"))
+    user_id: Mapped[int] = mapped_column(BigInteger, ForeignKey("users.id", ondelete="cascade", onupdate="cascade"))
     user: Mapped["User"] = relationship("User", back_populates="oauth_accounts")
