+++ app/profiles/models/contact.py
@@ -1,5 +1,5 @@
 from typing import TYPE_CHECKING
-from sqlalchemy import ForeignKey, Integer, String, UniqueConstraint
+from sqlalchemy import BigInteger, ForeignKey, String, UniqueConstraint
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from app.core.db.base_model import BaseModel, DateMixin
@@ -11,7 +11,7 @@ if TYPE_CHECKING:
 class Contact(BaseModel, DateMixin):
     __tablename__ = "contacts"
 
-    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)  
+    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)  
     provider: Mapped[str] = mapped_column(String(30))
     contact: Mapped[str] = mapped_column(String(256))
